\name{loglik2}
\alias{loglik2}
\title{
Computes the survival log-likelihood
}
\description{
  Computes the survival log-likelihood
}
\usage{
loglik2(theta, interimData)
}
\arguments{
  \item{theta}{
    the three-element vector of \eqn{(\alpha, \beta, \gamma)}
}
  \item{interimData}{
    The interim data
}
}
\details{
  Computes the survival log-likelihood
}
\value{
  the survival log-likelihood
}
\references{
  Lai, Tze Leung and Lavori, Philip W. and Shih, Mei-Chiung. Sequential
  Design of Phase II-III Cancer Trials, Statistics in Medicine, Volume 31, issue 18, p.1944-1960, 2012.
}
\author{
  Mei-Chiung Shih, Balasubramanian Narasimhan, Pei He
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{design}
